+++ src/routes/(app)/students/[id]/edit/+page.server.ts
@@ -1,5 +1,5 @@
 import { fetchUser, updateUser } from '$lib/actions/user.actions';
-import { fail, superValidate } from 'sveltekit-superforms';
+import { fail, message, superValidate } from 'sveltekit-superforms';
 import { zod } from 'sveltekit-superforms/adapters';
 import type { Actions, PageServerLoad } from './$types';
 import { studentUpdateSchema } from './studentUpdate.schema';
@@ -9,7 +9,12 @@ import { fetchMajors } from '$lib/actions/major.actions';
 import { fetchMinors } from '$lib/actions/minor.actions';
 import { redirect } from '@sveltejs/kit';
 
-export const load: PageServerLoad = async ({ params }) => {
+export const load: PageServerLoad = async ({ params, locals }) => {
+	const role = locals.user?.role;
+
+	if (role === 'STUDENT') {
+		redirect(303, '/');
+	}
 	const { id } = params;
 
 	const student = await fetchStudent(id);
@@ -26,18 +31,32 @@ export const load: PageServerLoad = async ({ params }) => {
 export const actions: Actions = {
 	update: async (event) => {
 		const form = await superValidate(event, zod(studentUpdateSchema));
+		if (event.locals.user?.role !== 'ADMIN') {
+			return message(
+				form,
+				{ message: 'You do not have permission to edit students', type: 'failure' },
+				{ status: 403 }
+			);
+		}
 		if (!form.valid) return fail(404, { form });
 		console.log(form.data);
 
 		try {
-			const userId = await updateStudent(form.data, form.data.advisors);
+			await updateStudent(form.data, form.data.advisors);
 		} catch (err) {
 			console.error(err);
-			return fail(500, { message: 'An error occurred' });
+			return message(
+				form,
+				{ message: 'Failed to update student', type: 'failure' },
+				{ status: 400 }
+			);
 		}
-		return { form };
+		return message(form, { message: 'Student created', type: 'success' });
 	},
-	delete: async ({ params }) => {
+	delete: async ({ params, locals }) => {
+		if (locals.user?.role !== 'ADMIN') {
+			return fail(403, { message: 'You do not have permission to delete students' });
+		}
 		const { id } = params;
 		try {
 			await deleteStudent(id);
@@ -45,6 +64,6 @@ export const actions: Actions = {
 			console.error(err);
 			return fail(500, { message: 'Failed to delete student' });
 		}
-		return redirect(302, '/students');
+		return { success: true };
 	}
 };
